drivers: clock: rcar: Rename global includes file
Rename r-car clock driver global include file
in order to match other files names.
