+++ drivers/clock_control/clock_control_r8a7795_cpg_mssr.c
@@ -11,7 +11,7 @@
 #include <errno.h>
 #include <soc.h>
 #include <zephyr/drivers/clock_control.h>
-#include <zephyr/drivers/clock_control/rcar_clock_control.h>
+#include <zephyr/drivers/clock_control/renesas_cpg_mssr.h>
 #include <zephyr/dt-bindings/clock/renesas_cpg_mssr.h>
 #include <zephyr/dt-bindings/clock/r8a7795_cpg_mssr.h>
 #include "clock_control_renesas_cpg_mssr.h"
