+++ drivers/timer/rcar_cmt_timer.c
@@ -8,7 +8,7 @@
 #include <soc.h>
 #include <zephyr/drivers/timer/system_timer.h>
 #include <zephyr/drivers/clock_control.h>
-#include <zephyr/drivers/clock_control/rcar_clock_control.h>
+#include <zephyr/drivers/clock_control/renesas_cpg_mssr.h>
 
 #define DT_DRV_COMPAT renesas_rcar_cmt
